requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
anthropic==0.52.0
python-dotenv==1.2.2
discord.py==2.4.0
